requirements.txt
python-gitlab==5.0.0